# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Vehiculr API",
    "version": "1.0.0",
    "description": "Full API specification for the Vehiculr backend (Express + Mongoose)."
  },
  "servers": [
    { "url": "http://localhost:9002/api", "description": "Local" },
    { "url": "https://api.yourdomain.com/api", "description": "Production" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ApiError": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "message": { "type": "string", "example": "Error message" },
          "error": { "type": "string", "example": "Validation failed" }
        }
      },
...
      "Partner": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "fullName": { "type": "string" },
          "businessName": { "type": "string" },
          "phone": { "type": "string" },
          "expertise": { "type": "array", "items": { "type": "string" } },
          "shopLocation": {
            "type": "object",
            "properties": {
              "type": { "type": "string", "example": "Point" },
              "coordinates": { "type": "array", "items": { "type": "number" }, "example": [77.5946,12.9716] }
            }
          },
          "distanceInKm": { "type": "number", "example": 1.23 }
        }
      },
...
    }
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/partners/nearby": {
      "get": {
        "tags": ["Partners"],
        "summary": "Find nearby partners by lat/lng",
        "parameters": [
          { "name":"latitude","in":"query","schema":{"type":"number"} },
          { "name":"longitude","in":"query","schema":{"type":"number"} },
          { "name":"maxDistance","in":"query","schema":{"type":"number","description":"in kilometers"} }
        ],
        "responses": {
          "200": { "description": "Nearby partners", "content": { "application/json": { "schema": { "type":"object","properties":{ "success":{"type":"boolean"},"count":{"type":"integer"},"partners":{"type":"array","items":{"$ref":"#/components/schemas/Partner"}} } } } } },
          "400": { "description": "Invalid coords", "content": { "application/json": { "schema": { "$ref":"#/components/schemas/ApiError" } } } }
        }
      }
    },
...
    "/uploads": {
      "post": {
        "tags":["Uploads"],
        "summary":"Upload an image (multipart)",
        "security":[{"bearerAuth":[]}],
        "requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"file":{"type":"string","format":"binary"}}}}}},
        "responses":{"200":{"description":"Uploaded","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean"},"url":{"type":"string"}}}}}}}
      }
    }
  }
}